Use -i and -o for consistency with niv's tools
This changes the usage of the tool to be more consistent with niv's
tools. Input files are specified with -i and output files with -o. If an
output file is not specified, it will choose a correct filename based on
the input file. If the user does specify an output file, no additional
file format information is added (previously, .nwnt would be added even
if the path already had that file extension).